\name{hydrosanity-internal}
\alias{waterQuarters}
\alias{sqrtPalette}
\alias{sqrtPretty}
\alias{select.sites.BOM.AU}
\alias{setperiod_handler}
\alias{setregion_handler}
\alias{addToLog}
\alias{readGDAL_fixed}
\alias{prepanel.qqmath.fix}
\title{Internal Hydrosanity functions}
\description{Ignore these.}
\keyword{internal}

