\name{hydrosanity-package}
\alias{hydrosanity-package}
\alias{hydrosanity}
\docType{package}
\title{
Hydrosanity: graphical interface for exploring hydrological time series
}
\description{
See \code{help(package=hydrosanity)}.
}
\details{
\tabular{ll}{
Package: \tab hydrosanity\cr
Type: \tab Package\cr
Version: \tab \code{hydrosanity:::VERSION}\cr
Date: \tab 2007-03-24\cr
License: \tab GPL version 2 or newer\cr
}
}
\author{ Felix Andrews \email{felix@nfrac.org} }
\references{
coming soon...
}
\keyword{ package }
\seealso{
\code{\link{hydrosanity}}
}

