\name{hydrosanity}
\alias{hydrosanity}
\title{Start the Hydrosanity GUI}
\description{
  Start the Hydrosanity Graphical User Interface.
}
\usage{
hydrosanity(project = NULL)
}
\arguments{
  \item{project}{ Path to a hydrosanity project file to open. }
}
\details{
  ...should list available documentation here.
}
\references{
Andrews, F. (2007). Hydrosanity: a starting point for hydrological analysis.
To appear in \emph{Proceedings of MODSIM 2007 - International Congress on Modelling and Simulation, 
Christchurch, New Zealand, 10-13 December 2007}. Modelling and Simulation Society of Australia and New Zealand Inc.

Package home page: \url{http://hydrosanity.googlecode.com/}
}
\author{ Felix Andrews \email{felix@nfrac.org} }
\examples{
\dontrun{
hydrosanity()

# view the introductory paper
RShowDoc("Andrews_2007_Hydrosanity", package="hydrosanity")

}
}
\keyword{ environment }

