\name{hydrosanity-internal}
\alias{.hs_on_notebook_switch_page}
\alias{.hs_on_menu_quit_activate}
\alias{.hs_on_menu_about_activate}
\alias{.hs_on_export_log_button_clicked}
\title{Interal Hydrosanity user interface callbacks}
\description{These are exported from the package so that the GUI can
  access the callbacks. They should otherwise be ignored.}
\author{\email{felix@nfrac.org}}
\keyword{internal}

